Move ScoreBoard drawing to another class
