+++ src/ScoreBoard.java
@@ -1,18 +1,8 @@
 import objectdraw.*;
-import java.awt.Color;
 
-public class ScoreBoard {
+public class ScoreBoard extends ScoreBoardDisplay {
 	
-	private static final int BOX_WIDTH = 135;
-	private static final int BOX_HEIGHT = 80;
-	private static final int FONT_SIZE = 30;
-	private static final String TITLE = "SCORE";
-	
-	private DrawingCanvas canvas;
-	private Location boardLocation;
-	private Text scoreDisplay;
-	
-	private Score score;
+	private int score = 0;
 	
 	/**
 	 * Constructs the score board.
@@ -21,68 +11,25 @@ public class ScoreBoard {
 	 * @param canvas  Canvas where the score board is drawn.
 	 */
 	public ScoreBoard(Location loc, DrawingCanvas canvas) {
-		boardLocation = loc;
-		this.canvas = canvas;
-		
-		drawScoreBoard();
-		initializeScore();
-	}
-	
-	/**
-	 * Updates the score.
-	 * 
-	 * @param numPoints  The number of points to add to the current score.
-	 */
-	public void addToScore(int numPoints) {
-		score.add(numPoints);
-		redrawScore();
+		super(loc, canvas);
+		reset();
 	}
 	
 	/**
 	 * Resets the score board.
 	 */
 	public void reset() {
-		score.reset();
-		redrawScore();
+		score = 0;
+		update(score);
 	}
 	
-	private void drawScoreBoard() {
-		drawBox();
-		drawTitle();
-		drawScore();	
-	}
-	
-	private void drawBox() {
-		new FilledRect(boardLocation, BOX_WIDTH, BOX_HEIGHT, canvas).setColor(Color2048.CELL_BG);
-	}
-	
-	private void drawTitle() {
-		Text title = new Text(TITLE, boardLocation, canvas);
-		title.setColor(Color.WHITE);
-		title.setFontSize(FONT_SIZE);
-		centerText(title);
-	}
-	
-	private void drawScore() {
-		scoreDisplay = new Text("", boardLocation.getX(), boardLocation.getY() + BOX_HEIGHT/2, canvas);
-		scoreDisplay.setColor(Color.WHITE);
-		scoreDisplay.setFontSize(FONT_SIZE);
-	}
-	
-	private void initializeScore() {
-		score = new Score();
-		redrawScore();
-		centerText(scoreDisplay);
-	}
-
-	private void redrawScore() {
-		scoreDisplay.setText(score.getScore());
-		centerText(scoreDisplay);
-	}
-	
-	private void centerText(Text text) {
-		double x = boardLocation.getX()+(BOX_WIDTH-text.getWidth())/2;
-		double y = text.getY();
-		text.moveTo(x, y);
+	/**
+	 * Updates the score.
+	 * 
+	 * @param points  The number of points to add to the current score.
+	 */
+	public void addToScore(int points) {
+		score += points;
+		update(score);
 	}
 }
